packages/justhtml/justhtml-1.4.0.tar.gz/justhtml-1.4.0/tests/justhtml-tests/empty_stack_edge_cases.dat

#data
<!><svg><th><title><n><select><td>
#errors
(1,2): incorrectly-opened-comment
(1,4): expected-doctype-but-got-start-tag
(1,34): unexpected-start-tag-in-select

#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg th>
|         <svg title>
|           <n>
|             <select>
|   <td>


#data
<svg><slot><select/e><title/a><select><select></slot></select>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,20): unexpected-character-after-solidus-in-tag
(1,29): unexpected-character-after-solidus-in-tag
(1,46): unexpected-select-in-select
(1,53): unexpected-end-tag
(1,62): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg slot>
|         <svg select>
|           e=""
|           <svg title>
|             a=""
|             <select>


#data
<svg><tr><title><section
><math><p><tr>>
#errors
(1,5): expected-doctype-but-got-start-tag
(2,10): unexpected-html-element-in-foreign-content
(2,15): unexpected-character-implies-table-voodoo

#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg tr>
|         <svg title>
|           <section>
|             <math math>
|             <p>
|   <tr>
|     ">"
